divided controller into 2 classes divided UI into classes with more specific function unable to decide identity of UserChoiceBuilder.java. hope to address in report
