+++ src/main/java/com/csc207/cli/SchedulingController.java
@@ -2,12 +2,14 @@ package com.csc207.cli;
 
 import com.csc207.domain.*;
 
+import java.util.Scanner;
+
 /**
  * This class is responsible for initiating scheduling and putting of tasks and projects into weeks and checking if a
  * task or project can be scheduled in a week.
  */
 
-public class Controller {
+public class SchedulingController {
 
     /**
      * Initiates scheduling and putting of a fixed task if user chooses
@@ -40,31 +42,3 @@ public class Controller {
         Putter.putProject(projectTasksToPut[0].getName(), week, projectTasksToPut);
     }
 
-    /**
-     * Check if a non-fixed task can be scheduled
-     * @param week: the week to schedule into
-     * @param taskToSchedule: the task to check
-     */
-    public static boolean checkNonFixedTaskScheduling(Week week,NonFixedTask taskToSchedule){
-        return Checker.CheckScheduleNonFixedTask(week, taskToSchedule);
-    }
-
-    /**
-     * Check if a project task can be scheduled
-     * @param week: the week to schedule into
-     * @param projectToSchedule: the project check
-     */
-    public static boolean checkProjectScheduling(Week week,NonFixedTask[] projectToSchedule){
-        return Checker.CheckScheduleProject(week, projectToSchedule);
-    }
-
-    /**
-     * Check if a fixed task can be scheduled
-     * @param week: the week to schedule into
-     * @param taskToSchedule: the task to check
-     */
-    public static boolean checkFixedTaskScheduling(Week week,FixedTask taskToSchedule){
-        return Checker.CheckScheduleFixedTask(week, taskToSchedule);
-    }
-
-}
